+++ ios/Classes/NordicDfuPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface NordicDfuPlugin : NSObject<FlutterPlugin>
-@end