+++ lib/nordic_dfu.dart
@@ -1,3 +1,6 @@
-export 'package:nordic_dfu/src/android_special_paramter.dart';
-export 'package:nordic_dfu/src/ios_special_parameter.dart';
+export 'package:nordic_dfu/src/dfu_event_handler.dart';
 export 'package:nordic_dfu/src/nordic_dfu.dart';
+export 'package:nordic_dfu/src/parameters/android_parameters.dart';
+export 'package:nordic_dfu/src/parameters/android_special_parameter.dart';
+export 'package:nordic_dfu/src/parameters/darwin_parameters.dart';
+export 'package:nordic_dfu/src/parameters/ios_special_parameter.dart';